split the file `watch_transaction_for_confirmations` to ease testing
